ingestion/requirements.txt
```text
snowflake-snowpark-python